Fixed comment about Utils class
